bring in more "demand" tests from "prefer"
I changed the "demand" tests to use the same default fixed time tests as
the "prefer" tests.  I brought in the tests from "prefer" into "demand,"
knowing that they should either pass or fail with a good reason,
demonstrating the difference between demand and prefer.

This demonstrates what seems to be a shortcoming in the implementation.

It's Black Friday, November 24, 2006, about 1:00.  I ask for "morning"
and I get Saturday morning.  This is no good: morning is already in the
future, there's no need to skip a day.  Similarly, 12:56:06pm gets me
that time on *Saturday*.  I think this is definitely a bug.  If it's
9:00 and I ask for a future noon, I expect the one in three hours.

I ask for "Friday" and I get next Friday, December 1.  I don't know if
this is a bug or not.